SE-16 COMPLETE

+++ tm-server/src/main/java/ru/iokhin/tm/service/ProjectService.java
@@ -1,6 +1,7 @@
 package ru.iokhin.tm.service;
 
 import lombok.SneakyThrows;
+import org.apache.deltaspike.jpa.api.transaction.Transactional;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import ru.iokhin.tm.entityDTO.ProjectDTO;
@@ -10,27 +11,33 @@ import ru.iokhin.tm.api.service.IProjectService;
 import ru.iokhin.tm.entity.Project;
 import ru.iokhin.tm.entity.User;
 import ru.iokhin.tm.enumerated.Status;
-import ru.iokhin.tm.repository.ProjectRepository;
-import ru.iokhin.tm.repository.UserRepository;
+import ru.iokhin.tm.exeption.AuthException;
 import ru.iokhin.tm.util.StringValidator;
 
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
+import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @ApplicationScoped
 public class ProjectService extends AbstractService<ProjectDTO> implements IProjectService {
 
+    @NotNull
+    private final IProjectRepository projectRepository;
+
+    @NotNull
+    private final IUserRepository userRepository;
+
     @Inject
-    public ProjectService(@NotNull EntityManagerFactory factory) {
-        super(factory);
+    public ProjectService(IProjectRepository projectRepository, IUserRepository userRepository) {
+        this.projectRepository = projectRepository;
+        this.userRepository = userRepository;
     }
 
     @Override
-    public ProjectDTO add(@NotNull final String userId, @NotNull final String name) {
+    @Transactional
+    public ProjectDTO add(@NotNull final String userId, @NotNull final String name) throws AuthException {
         StringValidator.validate(name);
         ProjectDTO project = new ProjectDTO(userId, name);
         merge(project);
@@ -38,90 +45,110 @@ public class ProjectService extends AbstractService<ProjectDTO> implements IProj
     }
 
     @Override
-    public ProjectDTO edit(@NotNull final String userId, @NotNull final String id, @NotNull final String name) {
+    @Transactional
+    public ProjectDTO edit(@NotNull final String userId, @NotNull final String id, @NotNull final String name) throws AuthException {
         StringValidator.validate(name, id, userId);
-        @Nullable final ProjectDTO project = findOneByUserId(userId, id);
+        @Nullable final Project project = projectRepository.findAnyByUserAndId(getUser(userId), id);
         if (project == null) return null;
         project.setName(name);
-        merge(project);
-        return project;
+        projectRepository.save(project);
+        return project.getProjectDTO();
     }
 
     @Override
-    public ProjectDTO edit(@NotNull String userId, @NotNull String id, @NotNull String name, @NotNull Status status) {
+    @Transactional
+    public ProjectDTO edit(@NotNull String userId, @NotNull String id, @NotNull String name, @NotNull Status status)
+            throws AuthException {
         StringValidator.validate(name, id, userId);
-        @Nullable final ProjectDTO project = findOneByUserId(userId, id);
-        if (project == null) return null;
+        @Nullable final Project project = projectRepository.findAnyByUserAndId(getUser(userId), id);
+        if (project == null) throw new AuthException("PROJECT NOT FOUND");
         project.setName(name);
         project.setStatus(status);
-        merge(project);
-        return project;
+        projectRepository.save(project);
+        return project.getProjectDTO();
     }
 
     @Override
-    public ProjectDTO removeByUserId(@NotNull final String userId, @NotNull final String id) {
+    @Transactional
+    public ProjectDTO removeByUserId(@NotNull final String userId, @NotNull final String id) throws AuthException {
         StringValidator.validate(id, userId);
-        @Nullable final ProjectDTO project = findOneByUserId(userId, id);
-        if (project == null) return null;
-        remove(project);
-        return project;
+        @Nullable final Project project = projectRepository.findAnyByUserAndId(getUser(userId), id);
+        if (project == null) throw new AuthException("PROJECT NOT FOUND");
+        @NotNull final ProjectDTO projectDTO = project.getProjectDTO();
+        remove(projectDTO);
+        return projectDTO;
     }
 
     @Override
     @SneakyThrows
+    @Transactional
     public void removeAllByUserId(@NotNull final String userId) {
         StringValidator.validate(userId);
-        @NotNull final List<ProjectDTO> projects = findAllByUserId(userId);
-        projects.forEach(this::remove);
+        User user = getUser(userId);
+        if (user == null) throw new AuthException("USER NOT FOUND");
+        @NotNull final List<Project> projects = projectRepository.findByUser(user);
+        if (projects == null) throw new AuthException("PROJECTS NOT FOUND");
+        for (Project project : projects) {
+            projectRepository.remove(project);
+        }
     }
 
     @Override
-    @SneakyThrows
-    public List<ProjectDTO> findAllByUserId(@NotNull final String userId) {
+    @Transactional(readOnly = true)
+    public List<ProjectDTO> findAllByUserId(@NotNull final String userId) throws AuthException {
         StringValidator.validate(userId);
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @NotNull final List<ProjectDTO> projects = projectRepository.findAllByUserId(getUser(userId, em))
-                .stream().map(Project::getProjectDTO)
+        @Nullable final User user = userRepository.findBy(userId);
+        if (user == null) throw new AuthException("USER NOT FOUND");
+        @Nullable final List<Project> projects = projectRepository.findByUser(user);
+        if (projects == null) throw new AuthException("PROJECTS NOT FOUND");
+        return projects.stream().map(Project::getProjectDTO)
                 .collect(Collectors.toList());
-        return projects;
     }
 
     @Override
-    public ProjectDTO findOneByUserId(@NotNull final String userId, @NotNull final String id) {
+    @Transactional(readOnly = true)
+    public ProjectDTO findOneByUserId(@NotNull final String userId, @NotNull final String id) throws AuthException {
         StringValidator.validate(userId, id);
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @Nullable final Project project = projectRepository.findOneByUserId(getUser(userId, em), id);
+        @Nullable final User user = userRepository.findBy(userId);
+        if (user == null) throw new AuthException("USER NOT FOUND");
+        @Nullable final Project project = projectRepository.findAnyByUserAndId(getUser(userId), id);
         if (project == null) return null;
         return project.getProjectDTO();
     }
 
     @Override
     @SneakyThrows
-    public List<ProjectDTO> sortByUserId(@NotNull final String userId, @NotNull final String parameter) {
+    @Transactional(readOnly = true)
+    public List<ProjectDTO> sortByUserId(@NotNull String userId, @NotNull String parameter) {
         StringValidator.validate(userId, parameter);
-        if ("order".equals(parameter)) return findAllByUserId(userId);
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @Nullable final List<ProjectDTO> projects = projectRepository.sortByUserId(getUser(userId, em), parameter)
-                .stream().map(Project::getProjectDTO).collect(Collectors.toList());
-        return projects;
+        @Nullable final User user = getUser(userId);
+        if (user == null) return null;
+        switch (parameter) {
+            case "order":
+                return findAllByUserId(userId);
+            case "status":
+                return projectRepository.sortByStatus(user).stream().map(Project::getProjectDTO).collect(Collectors.toList());
+            case "dateStart":
+                return projectRepository.sortByDateStart(user).stream().map(Project::getProjectDTO).collect(Collectors.toList());
+            case "dateEnd":
+                return projectRepository.sortByDateEnd(user).stream().map(Project::getProjectDTO).collect(Collectors.toList());
+            default:
+                throw new IllegalArgumentException("WRONG PARAMETER. VALID PARAMETERS: order, status, dateStart, dateEnd");
+        }
     }
 
     @Override
     @SneakyThrows
+    @Transactional(readOnly = true)
     public List<ProjectDTO> findByPartOfNameOrDescription(@NotNull final String userId, @NotNull final String part) {
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @Nullable final List<ProjectDTO> projects = projectRepository
-                .findByPartOfNameOrDescription(getUser(userId, em), part)
-                .stream().map(Project::getProjectDTO).collect(Collectors.toList());
-        return projects;
+        @Nullable final List<Project> projects = projectRepository.findByPartOfNameOrDescription(getUser(userId), part);
+        if (projects == null) throw new AuthException("PROJECTS NOT FOUND");
+        return projects.stream().map(Project::getProjectDTO).collect(Collectors.toList());
     }
 
     @Override
-    public Project getProjectFromDTO(ProjectDTO projectDTO, EntityManager em) {
+    @Transactional
+    public Project getProjectFromDTO(@NotNull final ProjectDTO projectDTO) throws AuthException {
         Project project = new Project();
         project.setId(projectDTO.getId());
         project.setName(projectDTO.getName());
@@ -129,73 +156,45 @@ public class ProjectService extends AbstractService<ProjectDTO> implements IProj
         project.setStatus(projectDTO.getStatus());
         project.setDateStart(projectDTO.getStartDate());
         project.setDateEnd(projectDTO.getEndDate());
-        project.setUser(getUser(projectDTO.getParentId(), em));
+        project.setUser(getUser(projectDTO.getParentId()));
         return project;
     }
 
     @Override
-    public User getUser(@NotNull final String userId, @NotNull final EntityManager em) {
-        @NotNull final IUserRepository userRepository = new UserRepository(em);
-        @NotNull final User user = userRepository.findOne(userId);
+    @Transactional(readOnly = true)
+    public User getUser(@NotNull final String userId) throws AuthException {
+        @Nullable final User user = userRepository.findBy(userId);
+        if (user == null) throw new AuthException("USER IS NULL");
         return user;
     }
 
     @Override
-    public void persist(@NotNull ProjectDTO projectDTO) {
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @NotNull final Project project = getProjectFromDTO(projectDTO, em);
-        try {
-            em.getTransaction().begin();
-            projectRepository.persist(project);
-            em.getTransaction().commit();
-        } catch (Exception e) {
-            em.getTransaction().rollback();
-            throw e;
-        } finally {
-            em.close();
-        }
+    @Transactional
+    public void persist(@NotNull final ProjectDTO projectDTO) throws AuthException {
+        @NotNull final Project project = getProjectFromDTO(projectDTO);
+        projectRepository.save(project);
     }
 
     @Override
-    public void merge(@NotNull ProjectDTO projectDTO) {
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @NotNull final Project project = getProjectFromDTO(projectDTO, em);
-        try {
-            em.getTransaction().begin();
-            projectRepository.merge(project);
-            em.getTransaction().commit();
-        } catch (Exception e) {
-            em.getTransaction().rollback();
-            throw e;
-        } finally {
-            em.close();
-        }
+    @Transactional
+    public void merge(@NotNull final ProjectDTO projectDTO) throws AuthException {
+        @NotNull final Project project = getProjectFromDTO(projectDTO);
+        projectRepository.save(project);
     }
 
     @Override
-    public ProjectDTO findOne(@NotNull String id) {
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @Nullable final Project project = projectRepository.findOne(id);
+    @Transactional(readOnly = true)
+    public ProjectDTO findOne(@NotNull final String id) {
+        @Nullable final Project project = projectRepository.findBy(id);
+        if (project == null) return null;
         return project.getProjectDTO();
     }
 
     @Override
-    public void remove(@NotNull ProjectDTO projectDTO) {
-        @NotNull final EntityManager em = factory.createEntityManager();
-        @NotNull final IProjectRepository projectRepository = new ProjectRepository(em);
-        @NotNull final Project project = getProjectFromDTO(projectDTO, em);
-        try {
-            em.getTransaction().begin();
-            projectRepository.remove(project);
-            em.getTransaction().commit();
-        } catch (Exception e) {
-            em.getTransaction().rollback();
-            throw e;
-        } finally {
-            em.close();
-        }
+    @Transactional
+    public void remove(@NotNull final ProjectDTO projectDTO) throws AuthException {
+        @Nullable final Project project = projectRepository.findBy(projectDTO.getId());
+        if (project == null) throw new AuthException("PROJECT IS NULL");
+        projectRepository.remove(project);
     }
 }
